Roll forward of https://github.com/google/error-prone/commit/65cf8acc727cb4dc1fcfd77d641ace7f869d5b4f: Push "lost type" logic up into AbstractReturnValueIgnored.
This isn't super-specific to Future, and is useful to reuse.

PiperOrigin-RevId: 339216378
